% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_unique_key}
\alias{check_unique_key}
\title{Check that data are uniquely identified by key columns}
\usage{
check_unique_key(data, keys)
}
\arguments{
\item{data}{A data frame.}

\item{keys}{Character vector of key column names.}
}
\value{
Invisibly, \code{data} unchanged if the keys are unique; otherwise an
error is raised.
}
\description{
Check that data are uniquely identified by key columns
}
\keyword{internal}
