% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_year_range}
\alias{check_year_range}
\title{Check year range}
\usage{
check_year_range(start_year, end_year, min_year = NULL, max_year = NULL)
}
\arguments{
\item{start_year}{First year.}

\item{end_year}{Last year.}

\item{min_year}{Optional. If given, \code{start_year} must be at least this.}

\item{max_year}{Optional. If given, \code{end_year} must be at most this.}
}
\value{
Invisibly \code{TRUE} if the range is valid; otherwise an error is
raised.
}
\description{
Check year range
}
\keyword{internal}
