% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cn_bib}
\alias{cn_bib}
\title{Bundled bibliography for cn_cite()}
\format{
A \code{bibentry} object of length 22.
}
\source{
\code{data-raw/contentiousR_bib.bib} and \code{data-raw/contentiousR_bib.R}.
}
\usage{
cn_bib
}
\description{
A \code{bibentry} object with one entry per bundled or interoperable data
source, used by \code{\link[=cn_cite]{cn_cite()}}. Each entry's \code{keywords} field lists the
\code{contentiousR} function(s) and/or \verb{dataset =} value(s) it documents.
}
\keyword{datasets}
