% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_gdp}
\alias{add_gdp}
\title{Add GDP data to a state panel}
\usage{
add_gdp(panel, dataset = c("gapminder", "fariss"))
}
\arguments{
\item{panel}{A data frame produced by \code{build_states_panel()}, containing
a \code{cow} or \code{gw} column and a \code{year} column.}

\item{dataset}{Dataset to use: \code{"gapminder"} (default) or \code{"fariss"}.
See \code{load_gdp_data()} for details and added columns.}
}
\value{
The input panel with GDP columns added via left join.
}
\description{
Joins GDP data onto an existing state panel. The coding system (\code{"cow"} or
\code{"gw"}) and year range are detected automatically from the panel.
}
\examples{
panel <- build_states_panel(1990, 2010, coding_system = "cow") |>
  add_gdp()

panel <- build_states_panel(1900, 2010, coding_system = "cow") |>
  add_gdp(dataset = "fariss")

}
\references{
Gapminder. (2024). \emph{GDP per capita in constant PPP dollars}.
\url{https://www.gapminder.org/gdp-per-capita/}

Fariss, C. J., Anders, T., Markowitz, J. N., & Barnum, M. (2022). New
estimates of over 500 years of historic GDP and population data.
\emph{Journal of Conflict Resolution}, 66(3), 553-591. \doi{10.1177/00220027211054432}
}
