% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_states_panel.R
\name{build_states_cow_panel}
\alias{build_states_cow_panel}
\title{Build a COW-coded state panel}
\usage{
build_states_cow_panel(
  start_year = 1946,
  end_year = 2013,
  exclude_microstates = TRUE,
  exclude_non_un = TRUE,
  exclude_islands = FALSE
)
}
\arguments{
\item{start_year}{First year. Must be \code{1816} or later.}

\item{end_year}{Last year. Must be \code{2025} or earlier.}

\item{exclude_microstates}{Logical. Exclude entities identified as
microstates by \code{\link[states:state_panel]{states::state_panel()}}.}

\item{exclude_non_un}{Logical. Exclude the package's fixed set of non-UN
entities (Kosovo, Taiwan, Hong Kong, the European Union, and South Vietnam
for COW; Kosovo, Taiwan, and South Vietnam for GW).}

\item{exclude_islands}{Logical. Exclude six small Caribbean island states:
Dominica, Grenada, Saint Lucia, Saint Vincent and the Grenadines, Antigua
and Barbuda, and Saint Kitts and Nevis.}
}
\value{
A data frame with COW-coded state-year observations.
}
\description{
Build a COW-coded state panel
}
\keyword{internal}
