% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_flag}
\alias{check_flag}
\title{Check a logical flag}
\usage{
check_flag(value, name)
}
\arguments{
\item{value}{Value to validate.}

\item{name}{Argument name used in an error message.}
}
\description{
Check a logical flag
}
\keyword{internal}
