% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_max}
\alias{safe_max}
\title{Return a typed missing value for an all-missing maximum}
\usage{
safe_max(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
Return a typed missing value for an all-missing maximum
}
\keyword{internal}
