% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_sum}
\alias{safe_sum}
\title{Sum a vector while preserving all-missing groups}
\usage{
safe_sum(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Sum a vector while preserving all-missing groups
}
\keyword{internal}
